fixed compiler warning

+++ MendeleySDK/MDLMendeleyAPIClient.m
@@ -460,7 +460,7 @@ requiresAuthentication:(BOOL)requiresAuthentication
     NSData *data = [NSData dataWithContentsOfURL:fileURL];
     uint8_t digest[CC_SHA1_DIGEST_LENGTH];
     
-    CC_SHA1(data.bytes, data.length, digest);
+    CC_SHA1(data.bytes,(unsigned int) data.length, digest);
     
     NSMutableString *output = [NSMutableString stringWithCapacity:CC_SHA1_DIGEST_LENGTH * 2];
